Remove global-interface
