+++ packages/viz/src/lib/sources/CARTOSource.ts
@@ -1,7 +1,12 @@
 import { Credentials, defaultCredentials } from '@carto/toolkit-core';
 import { MapInstance, MapOptions, Maps } from '@carto/toolkit-maps';
-import { Source, SourceProps, SourceMetadata } from './Source';
-import { NumericFieldStats, CategoryFieldStats } from '../global-interfaces';
+import {
+  Source,
+  SourceProps,
+  SourceMetadata,
+  NumericFieldStats,
+  CategoryFieldStats
+} from './Source';
 import { parseGeometryType } from '../style/helpers/utils';
 import { sourceErrorTypes, SourceError } from '../errors/source-error';
