+++ packages/viz/src/lib/style/helpers/utils.ts
@@ -1,7 +1,7 @@
 import { RGBAColor } from '@deck.gl/aggregation-layers/utils/color-utils';
 import { getColorPalette } from '../palettes';
-import { GeometryType } from '../../global-interfaces';
 import { Classifier } from '../../utils/Classifier';
+import { GeometryType } from '../../sources/Source';
 
 export function getUpdateTriggers(accessorFunction: Record<string, unknown>) {
   return {
